0.5.4 hot fix huix
теперь всё стреляет

+++ Assets/Weapons/Components/On Impact/ImpactDamageComponent.cs
@@ -8,7 +8,7 @@ public class ImpactDamageComponent : BaseImpactComponent
 
     public override void ProjectileImpact(GameObject other, Transform t)
     {
-        other.GetComponent<EnemyHealth>()?.TakeDamage((int)damage);
+        other.GetComponent<EnemyHealth>()?.TakeDamage(damage);
     }
 
     public override BaseWeaponComponent CloneComponent()
